requirements.txt
dagster~=1.11.2
dagster-webserver
psycopg2-binary~=2.9.10
sqlalchemy~=2.0.41
python-dotenv~=1.1.0
pandas~=2.2.3
python-dateutil~=2.9.0.post0
numpy~=2.2.6
matplotlib~=3.10.3
scipy~=1.15.3
reportlab~=4.4.1
psycopg==3.2.9
requests
Flask
Jinja2

weasyprint

playwright==1.54.0
PyPDF2==3.0.1

psycopg==3.2.9

psycopg==3.2.9

oracledb>=2.5,<3

openpyxl